move the error code mapping logic to vespa client
